+++ reeltok.api/reeltok.api.auth/Utils/DateTimeUtils.cs
@@ -2,12 +2,12 @@ namespace reeltok.api.auth.Utils
 {
     internal static class DateTimeUtils
     {
-        internal static uint DateTimeToUnixTime(DateTime timeToConvert)
+        internal static long DateTimeToUnixTime(DateTime timeToConvert)
         {
-            return (uint)new DateTimeOffset(timeToConvert).ToUnixTimeSeconds();
+            return new DateTimeOffset(timeToConvert).ToUnixTimeSeconds();
         }
 
-        internal static DateTime UnixTimeToDateTime(uint timeToConvert)
+        internal static DateTime UnixTimeToDateTime(long timeToConvert)
         {
             return DateTimeOffset.FromUnixTimeSeconds(timeToConvert).UtcDateTime;
         }
